GSDBE Google Sheets Database Editor V0.2.0
this is a google sheets editor program that utilizes my GS API, in order to allow the program from wherever it is loaded to edit in real time the contents of any GS file.

If you want to try it out, just make your own GS API using Google Cloud. Set your Gmail account as a test user, then download the JSON credentials from the Google Cloud Console. Keep all files in the same folder as it makes it easier to keep track of things. once all that is set the "sheets.txt" file is how the bulk access feature works. If you want to view multiple GS files at once with GSDBE then copy and paste what I have in mine then just copy over any and all Sheet IDs from the URL to the txt file. this is merely just for practice on my end so feel free to do anything you want with this.